moving superdroplet to a new file
